requirements.txt
gradio>=4.0.0
langchain-pinecone
langchain-huggingface
langchain-core
langchain-openai
python-dotenv
requests
beautifulsoup4
lxml
flask
rank_bm25
transformers
sentence-transformers
torch
numpy
scikit-learn
appwrite
aiohttp
pinecone-client 